latest addon structure and test setup
