--- app-addon/initializers/cordova-auth.js
@@ -1,13 +1,0 @@
-import Session from '../utils/session';
-
-export default {
-  name: 'session',
-
-  initialize: function(container, application) {
-    container.register('session:main', Session);
-
-    application.inject('route', 'session', 'session:main');
-    application.inject('controller', 'session', 'session:main');
-    application.inject('view', 'session', 'session:main');
-  }
-};
