+++ app/initializers/cordova-auth.js
@@ -1,0 +1,13 @@
+import Session from '../services/session';
+
+export default {
+  name: 'session',
+
+  initialize: function(container, application) {
+    container.register('session:main', Session);
+
+    application.inject('route', 'session', 'session:main');
+    application.inject('controller', 'session', 'session:main');
+    application.inject('view', 'session', 'session:main');
+  }
+};
