+++ index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  name: 'ember-cli-cordova-auth'
+};
