--- test-app/app/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>TestApp</title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    {{BASE_TAG}}
-
-    <link rel="stylesheet" href="assets/vendor.css">
-    <link rel="stylesheet" href="assets/test-app.css">
-  </head>
-  <body>
-    <script>
-      window.TestAppENV = {{ENV}};
-      window.EmberENV = window.TestAppENV.EmberENV;
-    </script>
-    <script src="assets/vendor.js"></script>
-    <script src="assets/test-app.js"></script>
-    <script>
-      window.TestApp = require('test-app/app')['default'].create(TestAppENV.APP);
-    </script>
-  </body>
-</html>
