--- test-app/app/router.js
@@ -1,10 +1,0 @@
-import Ember from 'ember';
-
-var Router = Ember.Router.extend({
-  location: TestAppENV.locationType
-});
-
-Router.map(function() {
-});
-
-export default Router;
