--- test-app/bower.json
@@ -1,17 +1,0 @@
-{
-  "name": "test-app",
-  "dependencies": {
-    "handlebars": "~1.3.0",
-    "jquery": "^1.11.1",
-    "qunit": "~1.12.0",
-    "ember-qunit": "~0.1.5",
-    "ember": "1.5.1",
-    "ember-resolver": "~0.1.1",
-    "loader": "stefanpenner/loader.js#1.0.0",
-    "ember-cli-shims": "stefanpenner/ember-cli-shims#0.0.2",
-    "ember-load-initializers": "stefanpenner/ember-load-initializers#0.0.2",
-    "ember-qunit-notifications": "^0.0.3",
-    "ember-cli-test-loader": "rjackson/ember-cli-test-loader#0.0.2",
-    "jquery-mockjax": "~1.5.3"
-  }
-}
