--- test-app/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "test-app",
-  "version": "0.0.0",
-  "private": true,
-  "directories": {
-    "doc": "doc",
-    "test": "test"
-  },
-  "scripts": {
-    "start": "ember server",
-    "build": "ember build",
-    "test": "ember test",
-    "postinstall": "./node_modules/bower/bin/bower install && npm link ../"
-  },
-  "repository": "https://github.com/stefanpenner/ember-cli",
-  "engines": {
-    "node": ">= 0.10.0"
-  },
-  "author": "",
-  "license": "MIT",
-  "devDependencies": {
-    "ember-cli": "0.0.39",
-    "originate": "0.1.5",
-    "broccoli-ember-hbs-template-compiler": "^1.5.0",
-    "express": "^4.1.1",
-    "body-parser": "^1.2.0",
-    "glob": "^3.2.9",
-    "ember-cli-ic-ajax": "0.1.1",
-    "ember-cli-ember-data": "0.1.0",
-    "broccoli-asset-rev": "0.0.11",
-    "bower": "^1.3.7"
-  },
-  "dependencies": {
-    "ember-cli-cordova-auth": "0.0.2"
-  }
-}
