--- test-app/testem.json
@@ -1,6 +1,0 @@
-{
-  "framework": "qunit",
-  "test_page": "tests/index.html",
-  "launch_in_ci": ["PhantomJS"],
-  "launch_in_dev": ["PhantomJS", "Chrome"]
-}
