--- test-app/tests/helpers/resolver.js
@@ -1,9 +1,0 @@
-import Resolver from 'ember/resolver';
-
-var resolver = Resolver.create();
-
-resolver.namespace = {
-  modulePrefix: 'test-app'
-};
-
-export default resolver;
