--- test-app/tests/helpers/start-app.js
@@ -1,30 +1,0 @@
-/* global require */
-
-var Application = require('test-app/app')['default'];
-var Router = require('test-app/router')['default'];
-import Ember from 'ember';
-
-export default function startApp(attrs) {
-  var App;
-
-  var attributes = Ember.merge({
-    // useful Test defaults
-    rootElement: '#ember-testing',
-    LOG_ACTIVE_GENERATION:false,
-    LOG_VIEW_LOOKUPS: false
-  }, attrs); // but you can override;
-
-  Router.reopen({
-    location: 'none'
-  });
-
-  Ember.run(function(){
-    App = Application.create(attributes);
-    App.setupForTesting();
-    App.injectTestHelpers();
-  });
-
-  App.reset(); // this shouldn't be needed, i want to be able to "start an app at a specific URL"
-
-  return App;
-}
