--- test-app/tests/index.html
@@ -1,50 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>TestApp Tests</title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    {{BASE_TAG}}
-
-    <link rel="stylesheet" href="assets/vendor.css">
-    <link rel="stylesheet" href="assets/test-app.css">
-    <link rel="stylesheet" href="assets/qunit.css">
-    <style>
-      #ember-testing-container {
-        position: absolute;
-        background: white;
-        bottom: 0;
-        right: 0;
-        width: 640px;
-        height: 384px;
-        overflow: auto;
-        z-index: 9999;
-        border: 1px solid #ccc;
-      }
-      #ember-testing {
-        zoom: 50%;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="qunit"></div>
-    <div id="qunit-fixture"></div>
-
-    <script>
-      window.TestAppENV = {{ENV}};
-      window.EmberENV = window.TestAppENV.EmberENV;
-    </script>
-    <script src="assets/qunit.js"></script>
-    <script src="assets/qunit-notifications.js"></script>
-    <script src="assets/vendor.js"></script>
-    <script src="assets/test-app.js"></script>
-    <script src="testem.js"></script>
-    <script>
-      require('test-app/tests/test-helper');
-    </script>
-    <script src="assets/test-loader.js"></script>
-  </body>
-</html>
