+++ tests/dummy/app/index.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <title>Dummy</title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    {{BASE_TAG}}
+
+    <link rel="stylesheet" href="assets/vendor.css">
+    <link rel="stylesheet" href="assets/dummy.css">
+  </head>
+  <body>
+    <script>
+      window.EmberENV = {{EMBER_ENV}};
+    </script>
+    <script src="assets/vendor.js"></script>
+    <script src="assets/dummy.js"></script>
+    <script>
+      window.Dummy = require('dummy/app')['default'].create({{APP_CONFIG}});
+    </script>
+  </body>
+</html>
