+++ tests/dummy/app/styles/app.css
