+++ tests/dummy/public/robots.txt
@@ -1,0 +1,3 @@
+# robotstxt.org/
+
+User-agent: *
