+++ tests/helpers/simple-module.js
