+++ tests/helpers/stub-request.js
